+++ hmm.h
@@ -1,10 +1,10 @@
-double logadd(double, double);
-double sum(double *, int);
-double forward(int *, int, int, int, double *, double *, double *);
-double forward_backward(int *, int, int, int, double *, double *, double *);
-void viterbi(int *, int, int, int, double *, double *, double *);
+float logadd(float, float);
+float sum(float *, int);
+float forward(int *, int, int, int, float *, float *, float *);
+float forward_backward(int *, int, int, int, float *, float *, float *);
+void viterbi(int *, int, int, int, float *, float *, float *);
 void baum_welch(int *data, int nseq, int iterations, int length, int nstates, int nobvs,
-        double *, double * , double *);
+        float *, float * , float *);
 void init_count();
 void update_prob();
 void usage();
